Use hide_console='hide-early'
Windows only. In console-enabled executable, hide or minimize the console window if the program owns the
console window (i.e., was not launched from existing console window). Depending on the setting, the
console is hidden/mininized either early in the bootloader execution ('hide-early', 'minimize-early') or
late in the bootloader execution ('hide-late', 'minimize-late'). The early option takes place as soon as
the PKG archive is found. In onefile builds, the late option takes place after application has unpacked
itself and before it launches the child process. In onedir builds, the late option takes place before
starting the embedded python interpreter.

https://github.com/pyinstaller/pyinstaller/blob/790d43f31cde74f3544ecf59f4c4d2035b2d5317/PyInstaller/building/api.py#L386